modular_code/source/requirements.txt
# pandas==1.3.0
# matplotlib==3.4.2
# tqdm==4.59.0
# opencv_python==4.5.3.56
# torch==1.9.0
# albumentations==0.1.12
# numpy==1.18.5
# PyYAML==6.0
# scikit_learn==1.0.2
pandas==1.3.0
matplotlib==3.4.2
tqdm==4.59.0
opencv_python==4.5.3.56
torch==1.9.0
albumentations==0.1.12
numpy==1.18.5
PyYAML==6.0
scikit_learn==1.0.2
